+++ lib/coinbase/user.rb
@@ -20,21 +20,14 @@ module Coinbase
     end
 
     # Creates a new Wallet belonging to the User.
+    # @param network_id [String] (Optional) the ID of the blockchain network. Defaults to 'base-sepolia'.
     # @return [Coinbase::Wallet] the new Wallet
-    def create_wallet
-      create_wallet_request = {
-        wallet: {
-          # TODO: Don't hardcode this.
-          network_id: 'base-sepolia'
-        }
-      }
-      opts = { create_wallet_request: create_wallet_request }
+    def create_wallet(create_wallet_options = {})
+      # For ruby 2.7 compatibility we cannot pass in keyword args when the create wallet
+      # options is empty
+      return Wallet.create if create_wallet_options.empty?
 
-      model = Coinbase.call_api do
-        wallets_api.create_wallet(opts)
-      end
-
-      Wallet.new(model)
+      Wallet.create(**create_wallet_options)
     end
 
     # Imports a Wallet belonging to the User.
@@ -44,14 +37,35 @@ module Coinbase
       Wallet.import(data)
     end
 
-    # Lists the IDs of the Wallets belonging to the User.
-    # @return [Array<String>] the IDs of the Wallets belonging to the User
-    def wallet_ids
-      wallets = Coinbase.call_api do
-        wallets_api.list_wallets
+    # Lists the Wallets belonging to the User.
+    # @param page_size [Integer] (Optional) the number of Wallets to return per page. Defaults to 10
+    # @param next_page_token [String] (Optional) the token for the next page of Wallets
+    # @return [Coinbase::Wallet] the Wallets belonging to the User
+    def wallets(page_size: 10, next_page_token: nil)
+      opts = {
+        limit: page_size
+      }
+
+      opts[:page] = next_page_token unless next_page_token.nil?
+
+      wallet_list = Coinbase.call_api do
+        wallets_api.list_wallets(opts)
       end
 
-      wallets.data.map(&:id)
+      # A map from wallet_id to address models.
+      address_model_map = {}
+
+      wallet_list.data.each do |wallet_model|
+        addresses_list = Coinbase.call_api do
+          addresses_api.list_addresses(wallet_model.id, { limit: Coinbase::Wallet::MAX_ADDRESSES })
+        end
+
+        address_model_map[wallet_model.id] = addresses_list.data
+      end
+
+      wallet_list.data.map do |wallet_model|
+        Wallet.new(wallet_model, seed: '', address_models: address_model_map[wallet_model.id])
+      end
     end
 
     # Saves a wallet to local file system. Wallet saved this way can be re-instantiated with load_wallets_from_local